Replace .source_file with .path in the .select to eliminate issue with scss files
Middleman's `sitemap.resources.source_file` lists the original `.scss` extension for scss files. If I have scss files in the css folder, the .select will not find them because of line 15: the name is appended with `".css"`.

All scss/sass/css.erb files compile to css so it makes sense to use the `.path` attribute which contains the file extensions in their final rendered format.